+++ src/components/ui/dialog.tsx
@@ -17,9 +17,9 @@ const DialogOverlay = React.forwardRef<
   }
 >(({ className, zIndex = "base", ...props }, ref) => {
   const zIndexMap = {
-    base: "z-40",
-    nested: "z-50",
-    alert: "z-[60]",
+    base: "z-[80]",
+    nested: "z-[90]",
+    alert: "z-[100]",
     top: "z-[110]",
   };
 
@@ -57,9 +57,9 @@ const DialogContent = React.forwardRef<
     ref,
   ) => {
     const zIndexMap = {
-      base: "z-40",
-      nested: "z-50",
-      alert: "z-[60]",
+      base: "z-[80]",
+      nested: "z-[90]",
+      alert: "z-[100]",
       top: "z-[110]",
     };